Extract text annotations panel
